formatting

+++ bookstore/src/main/java/com/mpolitakis/bookstore/repositories/UserRepository.java
@@ -1,19 +1,15 @@
 package com.mpolitakis.bookstore.repositories;
 
-
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mpolitakis.bookstore.models.User;
 
-
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
   User findByUsername(String username);
 
   Optional<User> findById(Long userId);
-
-  
 
 }
